renaming components
