+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import { Top } from "./components/top/";
+import { Top } from "./components/navbar-and-banner/";
 import { Portfolio } from "./components/portfolio/portfolio";
 import { Slider } from "./components/slider/slider";
 import { TextInfinite } from "./components/text-infinite";
